manage the disconnection of a user from web sockets (remove the user from the list of participants)
